# reminders/openapi.yaml
openapi: 3.0.3
info:
  title: Temporal Reminder API (minimal)
  version: 0.1.1
  description: |
    Minimal reminder management REST API built on Temporal Schedules.

    Endpoints operate on:
    - A Temporal Schedule: schedule_id (defaults to "rem-{reminder_id}")
    - A long-running record workflow: record_workflow_id (defaults to "record-{reminder_id}") of type "reminder.Record"
    - A fire workflow: type "reminder.Fire" started by the schedule action

    Schedule payload passthrough:
    - spec -> Temporal ScheduleSpec
    - policies -> Temporal SchedulePolicy
    - state -> Temporal ScheduleState

    Filtering is done via Temporal visibility query using Search Attributes on the record workflow.

servers:
  - url: http://localhost:8000

tags:
  - name: Reminders
  - name: Fires

paths:
  /reminders:
    post:
      tags: [Reminders]
      summary: Create a reminder
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
            examples:
              oneTimeCalendar:
                value:
                  title: "Test once (calendar)"
                  message: "fires at 2025-12-20T19:32:18Z"
                  tags: ["test"]
                  entity_type: "ticket"
                  entity_id: "TCK-123"
                  apprise_targets:
                    - "jsons://homeassistant.example/api/webhook/XYZ"
                  spec:
                    calendars:
                      - year: [{ start: 2025 }]
                        month: [{ start: 12 }]
                        day_of_month: [{ start: 20 }]
                        hour: [{ start: 19 }]
                        minute: [{ start: 32 }]
                        second: [{ start: 18 }]
                    time_zone_name: "UTC"
      responses:
        "200":
          description: Created reminder and schedule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReminderResponse'
        "422":
          description: FastAPI validation error (if body cannot be parsed as JSON object)

    get:
      tags: [Reminders]
      summary: List reminder record workflows
      operationId: listReminders
      parameters:
        - name: reminder_id
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by ReminderId Search Attribute:
            ReminderId = "{reminder_id}"
        - name: schedule_id
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by ReminderScheduleId Search Attribute:
            ReminderScheduleId = "{schedule_id}"
        - name: record_workflow_id
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by ReminderRecordWorkflowId Search Attribute:
            ReminderRecordWorkflowId = "{record_workflow_id}"

        - name: entity_type
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by generic association type using Temporal visibility:
            ReminderEntityType = "{entity_type}"
            Example: entity_type=ticket
        - name: entity_id
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by generic association id using Temporal visibility:
            ReminderEntityId = "{entity_id}"
            Example: entity_id=TCK-123

        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: |
            Back-compat single tag filter (contains):
            ReminderTags = "{tag}"

        - name: tag_any
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: |
            Match any of these tags (OR). Repeat query param:
            tag_any=a&tag_any=b -> (ReminderTags="a" OR ReminderTags="b")

        - name: tag_all
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: |
            Match all of these tags (AND). Repeat query param:
            tag_all=a&tag_all=b -> ReminderTags="a" AND ReminderTags="b"

        - name: target
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by Apprise target URL using Temporal visibility:
            ReminderTargets = "{target}"

        - name: q
          in: query
          required: false
          schema:
            type: string
          description: |
            Tokenized search; splits q into tokens and ANDs:
            ReminderText = "token" (Text SA equality matches whole tokens).

        - name: upcoming_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Filters by upcoming reminders:
            ReminderNextFireTime >= "{upcoming_after in UTC ISO8601}"

        - name: extra_filter
          in: query
          required: false
          schema:
            type: string
          description: |
            Raw Temporal visibility clause appended as AND ({extra_filter})

        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [next, start]
            default: next
          description: |
            API-side sort:
            - next: ReminderNextFireTime (nulls last)
            - start: start_time

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: |
            Max number of workflows returned (service caps at 1000).

      responses:
        "200":
          description: List of record workflows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowExecutionSummary'

  /reminders/{reminder_id}:
    get:
      tags: [Reminders]
      summary: Get a reminder record workflow by reminder_id
      operationId: getReminder
      parameters:
        - name: reminder_id
          in: path
          required: true
          schema:
            type: string
          description: Reminder id
      responses:
        "200":
          description: Reminder record workflow summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionSummary'
        "404":
          description: Not found

    put:
      tags: [Reminders]
      summary: Update a reminder (schedule + record workflow)
      operationId: updateReminder
      parameters:
        - name: reminder_id
          in: path
          required: true
          schema:
            type: string
          description: Reminder id (used to derive default schedule_id/record_workflow_id if not provided)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateReminderResponse'
        "422":
          description: FastAPI validation error (if body cannot be parsed as JSON object)

    delete:
      tags: [Reminders]
      summary: Delete a reminder (best-effort)
      operationId: deleteReminder
      parameters:
        - name: reminder_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted (best-effort; exceptions are swallowed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteReminderResponse'

  /fires:
    get:
      tags: [Fires]
      summary: List fire workflow executions
      operationId: listFires
      parameters:
        - name: scheduled_by_id
          in: query
          required: false
          schema:
            type: string
          description: |
            Filters by Temporal default Search Attribute:
            TemporalScheduledById = "{scheduled_by_id}"
        - name: after
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Filters by Temporal default Search Attribute:
            TemporalScheduledStartTime >= "{after in UTC ISO8601}"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of fire workflow executions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowExecutionSummary'

components:
  schemas:
    CreateReminderRequest:
      type: object
      additionalProperties: true
      description: |
        The service accepts an arbitrary JSON object and reads known keys.
        Unknown keys are ignored.

        Generic association:
        - entity_type/entity_id provide a generic external reference for querying via visibility.
        - A reminder may have at most one association (or none).
        - Multiple reminders may share the same entity_type/entity_id (1 entity -> many reminders).

        Filtering:
        - reminder_id, schedule_id, record_workflow_id, tags, entity_type/entity_id, apprise_targets, and next-fire are queryable via Search Attributes.
      properties:
        reminder_id:
          type: string
          description: Optional; if omitted the server generates a UUID.
        schedule_id:
          type: string
          description: Optional; default "rem-{reminder_id}"
        record_workflow_id:
          type: string
          description: Optional; default "record-{reminder_id}"

        entity_type:
          type: string
          description: Generic association type (e.g., "ticket", "order", "doc")
        entity_id:
          type: string
          description: Generic association id (e.g., "TCK-123")

        title:
          type: string
          default: ""
        message:
          type: string
          default: ""
        tags:
          type: array
          items: { type: string }
          default: []
        apprise_targets:
          type: array
          items: { type: string }
          default: []
        spec:
          $ref: '#/components/schemas/ScheduleSpec'
        policies:
          $ref: '#/components/schemas/SchedulePolicy'
        state:
          $ref: '#/components/schemas/ScheduleState'

    CreateReminderResponse:
      type: object
      required: [reminder_id, schedule_id, record_workflow_id, next_action_times]
      properties:
        reminder_id:
          type: string
        schedule_id:
          type: string
        record_workflow_id:
          type: string
        next_action_times:
          type: array
          items:
            type: string
            format: date-time
          description: Temporal schedule next_action_times returned by ScheduleHandle.describe()

    UpdateReminderRequest:
      type: object
      additionalProperties: true
      description: |
        The service accepts an arbitrary JSON object and reads known keys.
        Unknown keys are ignored.

        Behavior:
        - schedule_id and record_workflow_id default to rem-{reminder_id} and record-{reminder_id}.
        - title/message/tags/apprise_targets default to empty values if omitted.
        - spec/policies/state default to empty objects if omitted.
        - entity_type/entity_id are optional; clients may omit them to leave association unchanged.
          If explicitly provided as empty/omitted values, the association may be cleared.
      properties:
        schedule_id:
          type: string
        record_workflow_id:
          type: string

        entity_type:
          type: string
          description: Generic association type (e.g., "ticket", "order", "doc")
        entity_id:
          type: string
          description: Generic association id (e.g., "TCK-123")

        title:
          type: string
          default: ""
        message:
          type: string
          default: ""
        tags:
          type: array
          items: { type: string }
          default: []
        apprise_targets:
          type: array
          items: { type: string }
          default: []
        spec:
          $ref: '#/components/schemas/ScheduleSpec'
        policies:
          $ref: '#/components/schemas/SchedulePolicy'
        state:
          $ref: '#/components/schemas/ScheduleState'

    UpdateReminderResponse:
      type: object
      required: [ok, next_action_times]
      properties:
        ok:
          type: boolean
        next_action_times:
          type: array
          items:
            type: string
            format: date-time

    DeleteReminderResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    WorkflowExecutionSummary:
      type: object
      required: [workflow_id, run_id, type, status, start_time, search_attributes]
      properties:
        workflow_id:
          type: string
        run_id:
          type: string
        type:
          type: string
          description: Temporal workflow type (e.g., reminder.Record or reminder.Fire)
        status:
          type: string
        start_time:
          type: string
          format: date-time
        search_attributes:
          type: object
          additionalProperties: true
          description: |
            Raw search_attributes map returned by Temporal list_workflows for this execution.
            Note: values are often arrays; this API passes them through unchanged.

    ScheduleSpec:
      type: object
      additionalProperties: true
      description: |
        Passthrough for Temporal ScheduleSpec. This API structures the object using temporalio dataclasses.
        Common keys include calendars, intervals, cron_expressions, skip, start_at, end_at, jitter, time_zone_name.
      properties:
        calendars:
          type: array
          items: { $ref: '#/components/schemas/CalendarSpec' }
        intervals:
          type: array
          items: { $ref: '#/components/schemas/IntervalSpec' }
        cron_expressions:
          type: array
          items: { type: string }
        skip:
          type: array
          items: { $ref: '#/components/schemas/CalendarSpec' }
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        jitter:
          oneOf:
            - type: string
              description: ISO8601 duration (e.g., PT30S, PT5M)
            - type: number
              description: Seconds
        time_zone_name:
          type: string

    SchedulePolicy:
      type: object
      additionalProperties: true
      description: |
        Passthrough for Temporal SchedulePolicy (named "policies" in the API).
      properties:
        overlap:
          type: string
          description: Enum name of Temporal overlap policy (e.g., SKIP, BUFFER_ONE, etc.)
        catchup_window:
          oneOf:
            - type: string
              description: ISO8601 duration
            - type: number
              description: Seconds
        pause_on_failure:
          type: boolean

    ScheduleState:
      type: object
      additionalProperties: true
      description: |
        Passthrough for Temporal ScheduleState.
      properties:
        paused:
          type: boolean
        note:
          type: string
        limited_actions:
          type: boolean
        remaining_actions:
          type: integer

    IntervalSpec:
      type: object
      additionalProperties: true
      properties:
        every:
          oneOf:
            - type: string
              description: ISO8601 duration
            - type: number
              description: Seconds

    CalendarSpec:
      type: object
      additionalProperties: true
      properties:
        second:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        minute:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        hour:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        day_of_month:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        month:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        year:
          type: array
          items: { $ref: '#/components/schemas/Range' }
        day_of_week:
          type: array
          items: { $ref: '#/components/schemas/Range' }

    Range:
      type: object
      additionalProperties: true
      required: [start]
      properties:
        start: { type: integer }
        end: { type: integer }
        step: { type: integer }